pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "astronomical-events"
version = "0.1.0"
description = "Astronomical events notification from in-the-sky.org RSS feed"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.0",
    "python-telegram-bot>=20.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "apscheduler>=3.10.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
astronomical-events = "scripts.main:cli"

[tool.setuptools.packages.find]
include = ["src*"]
